+++ examples/full-screen/dummy-app2.py
@@ -1,0 +1,7 @@
+#!/usr/bin/env python3
+"""
+An empty full screen application without layout.
+"""
+from quo import Console
+
+Console().run()