+++ examples/full-screen/hello-world.py
@@ -3,24 +3,28 @@
 A simple example of a a text area displaying "Hello World!".
 """
 
-import quo
+from quo import container
+from quo import Console
+from quo.keys import KeyBinder
+from quo.layout import Layout
+from quo.widgets import Box, Frame, TextArea
 
 # Layout for displaying hello world.
 # (The frame creates the border, the box takes care of the margin/padding.)
-root_container = quo.widgets.Box(
-    quo.widgets.Frame(
-        quo.widgets.TextArea(
-            text="Hello world!\nPress control-c to quit.",
-            width=40,
-            height=10,
+root_container = Box(
+        Frame(
+            TextArea(
+                text="Hello world!\nPress control-c to quit.",
+                width=40,
+                height=10,
+                )
+            ),
         )
-    ),
-)
-layout = quo.layout.Layout(container=root_container)
+layout = Layout(container=root_container)
 
 
 # Key bindings.
-kb = quo.keys.KeyBinder()
+kb = KeyBinder()
 
 
 @kb.add("ctrl-c")
@@ -30,7 +34,7 @@ def _(event):
 
 
 # Build a main application object.
-application = quo.Console(
+application = Console(
         layout=layout,
         bind=kb, 
         full_screen=True