+++ examples/full-screen/no-layout.py
@@ -1,7 +1,0 @@
-#!/usr/bin/env python
-"""
-An empty full screen application without layout.
-"""
-from quo.console import Console
-
-Console(full_screen=True).run()