+++ examples/prompts/system-clipboard-integration.py
@@ -3,9 +3,11 @@
 Demonstration of a custom clipboard class.
 This requires the 'pyperclip' library to be installed.
 """
-import quo
+from quo import echo
+from quo.clipboard import PyperClipboard
+from quo.prompt import Prompt
 
-session = quo.Prompt()
+session = Prompt()
 
 if __name__ == "__main__":
     print("Emacs shortcuts:")
@@ -14,5 +16,5 @@ if __name__ == "__main__":
     print("    Press Control-W to cut to clipboard.")
     print("")
 
-    answer = session.prompt("Give me some input: ", clipboard=quo.clipboard.PyperClipboard())
-    quo.echo(f"You said: {answer}")
+    answer = session.prompt("Give me some input: ", clipboard=PyperClipboard())
+    echo(f"You said: {answer}")