+++ examples/prompts/toggle_password.py
@@ -3,23 +3,22 @@
 get_password function that displays asterisks instead of the actual characters.
 With the addition of a ControlT shortcut to hide/show the input.
 """
-import quo
+from quo import Condition
+from quo.prompt import Prompt
+from quo.keys import KeyBinder
 
-session = quo.Prompt()
-
+session = Prompt()
+kb = KeyBinder()
 def main():
     hidden = [True]  # Nonlocal
-    bindings = quo.keys.KeyBinder()
 
-    @bindings.add("ctrl-t")
+    @kb.add("ctrl-t")
     def _(event):
         "When ControlT has been pressed, toggle visibility."
         hidden[0] = not hidden[0]
 
     print("Type Control-T to toggle password visible.")
-    password = session.prompt(
-        "Password: ", is_password=quo.filters.Condition(lambda: hidden[0]), key_bindings=bindings
-    )
+    password = session.prompt( "Password: ", is_password=Condition(lambda: hidden[0]), bind=kb)
     print("You said: %s" % password)