+++ examples/types/float.py
@@ -1,0 +1,6 @@
+from quo.prompt import Prompt
+from quo.types import Float
+
+session = Prompt()
+
+float(session.prompt("Give me some input", type=Float))